requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.0.6
django-admin-interface==0.28.8
django-advanced-filters==2.0.0
django-colorfield==0.11.0
django-import-export==4.1.1
pillow==10.4.0
python-slugify==8.0.4
simplejson==3.19.2
sqlparse==0.5.0
tablib==3.5.0
text-unidecode==1.3
tzdata==2024.1
